+++ CHANGELOG.md
@@ -5,9 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[0.2.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/johnarp/proficiency-tracker/releases/tag/v0.0.1
+
+## [0.2.0] - 2026-03-14
+
+### Added
+
+- Settings page
+- Changes are saved in localStorage
+- Import/export saved data
+
+### Changed
+
+- Rank colors to better match the game
+
+### Fixed
+
+- Unreadable ranks and levels in pop-up when using dark mode
 
 ## [0.1.1] - 2026-03-13